Add OPC-UA plugin memory access utilities and refactor related components
- Introduced `opcua_memory.py` for direct memory access functions.
- Created `opcua_types.py` for type definitions including `VariableNode` and `VariableMetadata`.
- Added utility functions in `opcua_utils.py` for type mapping and value conversion.
- Refactored `opcua_plugin.py` to utilize new memory access and utility functions.
